requirements.txt
# Packages
langchain
openpyxl
streamlit
python-dotenv
pandas
numpy
chromadb
sentence-transformers
openai
anthropic
sqlalchemy